Welcome to E-Nguyen
Near 23rd and Classen, neon orange palm trees broadcast the way to the a magical warehouse of prosperity.   If you look towards the very boulevard where those seeking a better life built it, and you carry cash, you may find a sacred enclave as old as the beginning where you can refuel your now mortal vessel with the savory ambrosia of Neo Psygon.  The bones were simmered for hours to make the broth that now cooks the beef in front of you.  Rice noodles, bean spouts, basil & cilantro leaves, a dash of lime, and rooster sauce combine like Voltron and release an anti-gravitational wave that connects you to a network of our resonance from 89th & Penn to Valencia to the farthest bastions of HOPE and virtue.  In this state of fate-empowerment your karma will invoke wu wei into, through, and away from the impenetrable fog of futures unkind sold by the would-be disaffected tyrants & eager victims of a boring dystopia.  You may move in dimensions that we are said not to permit motion.  You will find that you are going there faster, pulling the fabric of space time until the reality stone shatters and all that is left is all that was and will be aspired to ad infinitum, forever, the future, the simple and undiluted river of the now, the freedom and the hands that build it.  We will take our prosperity to space, to where we define no boundaries, only waypoints.  We will not get there.  We will go there.  Welcome to E-Nguyen.