Use Conjugates
The FFT is actually a combination of two similar sets of data.  IIRC it's just sin and cosine but I need to look at the derivation again.  In any case, one is the conjugate and they are nearly mirror images of each other for our purposes.

One side effect is that we are also flipping the image in its current implementation, but this is an unintended side effect;  it's best to keep high frequencies in high texture addresses for consumers to not get sign-madness