Introducing Logorithmic Mapping
In a spectogram of 20Hz - 20kHz, the bounds of pretty typical human teenager hearing, a linear mapping leaves virtually all of the action huddled into one corner.

This PR introduces the conversion to a logorithmic space where the base is chosen so that each display bin is a constant ratio frequency higher than the previous.