Power Scaling 1.0
To go in the direction of handling the log/linear sampling mismatch, first, each linear bin is treated as though it is spread between the two neighboring bins centers.  This is the frequency envelope for the linear bin.

Second, the logarithmic frequency window, from one half power down to one half power higher, is "projected" onto these bins and, depending on how they overlap, calculates a fraction of the power to consider owned by this log bin.

This makes our sampling of the linear FFT aware of how much frequency space the linear bin represents vs how much the log bin overlaps it.  Low frequencies have up until now been severely over-represented because each log bin was acting like it had collected the power of the entire linear bin it sampled, making calibration of filtering a lost cause.

The highest frequencies add up several bins while the lowest frequencies stretch the bin out.

Left for a "good first issue" is adding a linear interpolation using a tent so that the linear bins will be smoothly blended.