! Introducing Failure
Fail is a convenient crate for writing fine-grained error types

The current state is being committed not because it is great, but because the existing implementation of the Error trait provides nothing approximating a foundation for ergonomic error handling.

Takeaways so far:

- We can get rid of the Box<dyn Error> type that is not going to ever result in helpful information being presented upstream while costing dynamic dispatch and heap allocation
- Rust Enums (tagged unions) are good for this.  We can nest these for hierarchy
- A hierarchy should exist since mass coercion is going to result in way too much plumbing and duplication
- The size achieved can be pretty small, mostly just a pointer to a static string and the union tag data
- Switching on Enum variants will probably be a pretty efficient way to utilize downstream information to recover or change behavior appopriately

Some errors are happening frequently and the lack of consistent logging behavior should be turned into something that works with the log flags.  Oh man, can't wait to write a logging function that classifies all error types into various levels of loudness again =)  Never did that before.